#249686 - Extends ParticipationCountService to check if participation is possible (replaces everything which is related to limit is reached checks) - Introduces MealGuestController (all guest related stuff is moved from MealController to this class) - Some cleanups in BaseController and Doorman. - CSS and JS fixes.

+++ src/Resources/js/modules/participant-counter.ts
@@ -27,7 +27,7 @@ export class ParticipantCounter {
         this.counter = parseInt(this.$counter.text().trim()) || 0;
 
         this.$limit = this.$participantsCountWrapper.find(this.limitSelector);
-        this.limit = parseInt(this.$limit.text().replace(this.delimiter, '').trim()) || 0;
+        this.limit = parseFloat(this.$limit.text().replace(this.delimiter, '').trim()) || 0;
 
         this.initState();
     }
